// theme/src/theme/types/theme.ts
export interface Theme {
  color: {
    primary: string;
    secondary: string;
    accentColor: string;
    default: string;
    light: string;
    transparent: string;
    border: string;
  };
  background: {
    default: string;
    primary: string;
    secondary: string;
    tertiary: string;
    quaternary: string;
    quinary: string;
    transparent: string;
    hover: string;
  };
  transitionOption: string;
  boxShadow: string;
}